Python program to split the file name.
This program helps you to split the file name. If proper file name is ensured on all files (files may be pdf,csv or whatever). This program will assist you to store the record in the database. By utilizing the this table you can populate on any simple web-form. Enjoy the day. Cheers